patches to make ompl compile in windows, thanks to nikos
--HG--
extra : convert_revision : f19d62bfdf148c821ab1010157daff6ee1412456

+++ trunk/src/ompl/control/planners/kpiece/src/KPIECE1.cpp
@@ -303,7 +303,7 @@ bool ompl::control::KPIECE1::selectMotion(Motion* &smotion, Grid::Cell* &scell)
 	content.reserve(tree_.grid.size());
 	tree_.grid.getContent(content);
 	for (std::vector<CellData*>::iterator it = content.begin() ; it != content.end() ; ++it)
-	    (*it)->score += 1.0 + log((*it)->iteration);
+	    (*it)->score += 1.0 + log((double)((*it)->iteration));
 	tree_.grid.updateAll();
     }
